requirements.txt
Flask
PyMySQL
openpyxl
python-docx
Babel
PyQt5
PyQtWebEngine
cx_Freeze
Frozen-Flask
flask
pymysql
openpyxl
python-docx
babel
PyQt5
PyQtWebEngine
pyinstaller
cryptography